separate reusable plugin registry with sorting
