+++ lib/inspec/fetcher.rb
@@ -3,31 +3,10 @@
 # author: Christoph Hartmann
 
 require 'inspec/plugins'
+require 'utils/plugin_registry'
 
 module Inspec
-  class Fetcher
-    @registry = {}
-
-    class << self
-      attr_reader :registry
-
-      def resolve(target)
-        modules.each do |m|
-          res = m.resolve(target)
-          return res unless res.nil?
-        end
-        nil
-      end
-
-      private
-
-      def modules
-        @registry.values
-                 .sort_by { |x| x.respond_to?(:priority) ? x.priority : 0 }
-                 .reverse
-      end
-    end
-  end
+  Fetcher = PluginRegistry.new
 
   def self.fetcher(version)
     if version != 1
